backend/requirements.txt
# RAiCare Backend Dependencies

# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database (MongoDB)
motor==3.3.2
beanie==1.24.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Image Upload (Cloudinary)
cloudinary==1.38.0

# AI Chatbot (Google Gemini)
google-genai==0.8.0

# Machine Learning (PyTorch)
torch>=2.6.0
torchvision>=0.16.0
Pillow>=10.0.0

# Email Validation
pydantic[email]==2.5.3

# Date/Time utilities
python-dateutil==2.8.2
